Issue #86 fix: 머지후 충돌 해결
