+++ app/src/main/java/com/lateinit/rightweight/data/RoutineApiService.kt
@@ -2,8 +2,7 @@ package com.lateinit.rightweight.data
 
 import com.lateinit.rightweight.data.database.mediator.SharedRoutineRequestBody
 import com.lateinit.rightweight.data.model.DocumentResponse
-import com.lateinit.rightweight.data.model.RoutineCollection
-import com.lateinit.rightweight.data.remote.model.SharedRoutineField
+import com.lateinit.rightweight.data.remote.model.*
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
@@ -22,22 +21,21 @@ interface RoutineApiService {
     suspend fun getSharedRoutines(
         @Body order: SharedRoutineRequestBody
     ): List<DocumentResponse<SharedRoutineField>>
-    suspend fun getRoutines(): DocumentResponse<RoutineCollection>
 
-    @POST("shared_routine")
+    @POST("documents/shared_routine")
     suspend fun shareRoutine(
         @Query("documentId") routineId: String,
         @Body rootField: RootField
     ): DocumentResponse<SharedRoutineField>
 
-    @POST("shared_routine/{routineId}/day")
+    @POST("documents/shared_routine/{routineId}/day")
     suspend fun shareRoutineDay(
         @Path("routineId") routineId: String,
         @Query("documentId") dayId: String,
         @Body rootField: RootField
     ): DocumentResponse<DayField>
 
-    @POST("shared_routine/{routineId}/day/{dayId}/exercise")
+    @POST("documents/shared_routine/{routineId}/day/{dayId}/exercise")
     suspend fun shareRoutineExercise(
         @Path("routineId") routineId: String,
         @Path("dayId") dayId: String,
@@ -45,7 +43,7 @@ interface RoutineApiService {
         @Body rootField: RootField
     ): DocumentResponse<ExerciseField>
 
-    @POST("shared_routine/{routineId}/day/{dayId}/exercise/{exerciseId}/exercise_set")
+    @POST("documents/shared_routine/{routineId}/day/{dayId}/exercise/{exerciseId}/exercise_set")
     suspend fun shareRoutineExerciseSet(
         @Path("routineId") routineId: String,
         @Path("dayId") dayId: String,
