+++ app/src/main/java/com/lateinit/rightweight/data/datasource/RoutineRemoteDataSourceImpl.kt
@@ -7,13 +7,14 @@ import com.lateinit.rightweight.data.RoutineApiService
 import com.lateinit.rightweight.data.database.AppDatabase
 import com.lateinit.rightweight.data.database.AppSharedPreferences
 import com.lateinit.rightweight.data.database.mediator.SharedRoutineRemoteMediator
+import com.lateinit.rightweight.data.remote.model.RootField
 import javax.inject.Inject
 
 class RoutineRemoteDataSourceImpl @Inject constructor(
     private val db: AppDatabase,
     private val api: RoutineApiService,
     private val appSharedPreferences: AppSharedPreferences
-): RoutineRemoteDataSource {
+) : RoutineRemoteDataSource {
 
 
     @OptIn(ExperimentalPagingApi::class)
@@ -25,21 +26,13 @@ class RoutineRemoteDataSourceImpl @Inject constructor(
     ) {
         db.sharedRoutineDao().getAllSharedRoutinesByPaging()
     }.flow
-import android.util.Log
-import com.google.gson.Gson
-import com.lateinit.rightweight.data.RoutineApiService
-import com.lateinit.rightweight.data.model.RootField
-import javax.inject.Inject
 
-class RoutineRemoteDataSourceImpl @Inject constructor(
-    private val routineApiService: RoutineApiService
-) : RoutineRemoteDataSource {
     override suspend fun shareRoutine(routineId: String, rootField: RootField) {
-        routineApiService.shareRoutine(routineId, rootField)
+        api.shareRoutine(routineId, rootField)
     }
 
     override suspend fun shareDay(routineId: String, dayId: String, rootField: RootField) {
-        routineApiService.shareRoutineDay(routineId, dayId, rootField)
+        api.shareRoutineDay(routineId, dayId, rootField)
     }
 
     override suspend fun shareExercise(
@@ -48,7 +41,7 @@ class RoutineRemoteDataSourceImpl @Inject constructor(
         exerciseId: String,
         rootField: RootField
     ) {
-        routineApiService.shareRoutineExercise(routineId, dayId, exerciseId, rootField)
+        api.shareRoutineExercise(routineId, dayId, exerciseId, rootField)
     }
 
     override suspend fun shareExerciseSet(
@@ -58,8 +51,7 @@ class RoutineRemoteDataSourceImpl @Inject constructor(
         exerciseSetId: String,
         rootField: RootField
     ) {
-        Log.d("RoutineExerciseAdapter", Gson().toJson(rootField))
-        routineApiService.shareRoutineExerciseSet(
+        api.shareRoutineExerciseSet(
             routineId,
             dayId,
             exerciseId,
@@ -67,4 +59,5 @@ class RoutineRemoteDataSourceImpl @Inject constructor(
             rootField
         )
     }
+
 }
