--- app/src/main/java/com/lateinit/rightweight/data/model/RootField.kt
@@ -1,8 +1,0 @@
-package com.lateinit.rightweight.data.model
-
-import com.google.gson.annotations.SerializedName
-
-data class RootField(
-    @SerializedName("fields")
-    val remoteData: RemoteData? = null
-)
