+++ app/src/main/java/com/lateinit/rightweight/data/remote/model/RootField.kt
@@ -3,6 +3,6 @@ package com.lateinit.rightweight.data.remote.model
 import com.google.gson.annotations.SerializedName
 
 data class RootField(
-    @field:SerializedName("fields")
+    @SerializedName("fields")
     val remoteData: RemoteData? = null
 )
