+++ app/src/main/java/com/lateinit/rightweight/data/repository/RoutineRepositoryImpl.kt
@@ -8,10 +8,10 @@ import com.lateinit.rightweight.data.database.intermediate.DayWithExercises
 import com.lateinit.rightweight.data.database.intermediate.RoutineWithDays
 import com.lateinit.rightweight.data.datasource.RoutineLocalDataSource
 import com.lateinit.rightweight.data.datasource.RoutineRemoteDataSource
-import com.lateinit.rightweight.data.model.DayField
-import com.lateinit.rightweight.data.model.ExerciseField
-import com.lateinit.rightweight.data.model.ExerciseSetField
-import com.lateinit.rightweight.data.model.RootField
+import com.lateinit.rightweight.data.remote.model.DayField
+import com.lateinit.rightweight.data.remote.model.ExerciseField
+import com.lateinit.rightweight.data.remote.model.ExerciseSetField
+import com.lateinit.rightweight.data.remote.model.RootField
 import com.lateinit.rightweight.util.toSharedRoutineField
 import javax.inject.Inject
